Allocate ID for each insertMsg
